Moved CommonFormEx and IFormView from Common/Controls to Common/SystemUtil
