+++ pwiz_tools/Shared/Common/SystemUtil/IFormView.cs
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-namespace pwiz.Common.Controls
+namespace pwiz.Common.SystemUtil
 {
     /// <summary>
     /// Marker interface for localization testing of forms
